pyproject.toml
[project]
name = "xui"
version = "0.1.0"
description = "A pygame GUI library"
dependencies = [
  "pygame-ce",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
